# frontend/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neuro Chess ML Engine",
    "description": "Asgardeo-protected API for chess move recommendations, RAG strategy analysis, and PDF move extraction.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/operations": {
      "get": {
        "summary": "Operational Status",
        "description": "Restricted operational counters; no uploaded documents or bearer tokens.",
        "operationId": "operational_status_api_v1_operations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/players": {
      "get": {
        "summary": "Available Players",
        "operationId": "available_players_api_v1_players_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/predict-move": {
      "post": {
        "summary": "Predict Move",
        "description": "Rank legal moves with a transparent heuristic and optional Redis cache.",
        "operationId": "predict_move_api_v1_predict_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovePredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovePredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ingest-async": {
      "post": {
        "summary": "Trigger Ingest",
        "description": "Triggers an asynchronous PGN ingestion task via RabbitMQ.",
        "operationId": "trigger_ingest_api_v1_ingest_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/predict-strategy": {
      "post": {
        "summary": "Predict Strategy",
        "description": "Generates an in-depth strategic analysis using Gemini and PostgreSQL (pgvector).",
        "operationId": "predict_strategy_api_v1_predict_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/predict-strategy/stream": {
      "post": {
        "summary": "Stream Strategy",
        "operationId": "stream_strategy_api_v1_predict_strategy_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/extract-page-moves": {
      "post": {
        "summary": "Extract Page Moves",
        "operationId": "extract_page_moves_api_v1_extract_page_moves_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_page_moves_api_v1_extract_page_moves_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ocr-jobs/{job_id}": {
      "get": {
        "summary": "Ocr Status",
        "operationId": "ocr_status_api_v1_ocr_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extract-page-image": {
      "post": {
        "summary": "Extract Page Image",
        "operationId": "extract_page_image_api_v1_extract_page_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_page_image_api_v1_extract_page_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Body_extract_page_image_api_v1_extract_page_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          }
        },
        "type": "object",
        "required": [
          "file",
          "page"
        ],
        "title": "Body_extract_page_image_api_v1_extract_page_image_post"
      },
      "Body_extract_page_moves_api_v1_extract_page_moves_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          }
        },
        "type": "object",
        "required": [
          "file",
          "page"
        ],
        "title": "Body_extract_page_moves_api_v1_extract_page_moves_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "IngestRequest"
      },
...
      "MovePredictionRequest": {
        "properties": {
          "fen": {
            "type": "string",
            "maxLength": 120,
            "minLength": 10,
            "title": "Fen"
          },
          "opponent_username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Username",
            "default": "Opponent"
          },
          "initial_fen": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Fen"
          },
          "moves": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 1500
              },
              {
                "type": "null"
              }
            ],
            "title": "Moves"
          }
        },
        "type": "object",
        "required": [
          "fen"
        ],
        "title": "MovePredictionRequest"
      },
...
      "PlayerSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          }
        },
        "type": "object",
        "required": [
          "name",
          "games"
        ],
        "title": "PlayerSummary"
      },
      "PlayersResponse": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "players"
        ],
        "title": "PlayersResponse"
      },
...
      "StrategyPredictionRequest": {
        "properties": {
          "opponent_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Opponent Name"
          },
          "context": {
            "type": "string",
            "maxLength": 2000,
            "title": "Context",
            "default": ""
          },
          "color": {
            "type": "string",
            "enum": [
              "any",
              "white",
              "black"
            ],
            "title": "Color",
            "default": "any"
          }
        },
        "type": "object",
        "required": [
          "opponent_name"
        ],
        "title": "StrategyPredictionRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}